Feat: 첫 번째 페이지 크롤링 구현 완료

+++ SJUDorm-backend/src/main/java/com/example/sjudormbackend/domain/FacilityPetitionLetter.java
@@ -10,7 +10,7 @@ import java.util.Date;
 public class FacilityPetitionLetter {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "facility_petition_letter_id")
-    private int facilityPetitionLetterId;
+    private Long facilityPetitionLetterId;
 
     @Column(name = "status")
     private boolean status;
